+++ src/main/java/com/inalogy/midpoint/connectors/ssh/utils/Constants.java
@@ -20,7 +20,6 @@ public class Constants {
     public static final String SPECIAL_CONNID_UID = "__UID__";
     public static final String SPECIAL_CONNID_PASSWORD = "__PASSWORD__";
 
-    public static final int SSH_RESPONSE_TIMEOUT = 15;
     /**
      * How often SshClient send keep alive packet
      */