+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.3] - 2024-05-15
+### Added
+
+- SSH_RESPONSE_TIMEOUT as configuration property
+- proper Message properties for configuration properties
+
 ## [1.2.2] - 2024-04-01
 ### Fixed
 
@@ -20,7 +26,7 @@ All notable changes to this project will be documented in this file.
 
 - public release
 
-
+[1.2.3]: https://github.com/inalogy/ssh-connector/releases/tag/v1.2.3
 [1.2.2]: https://github.com/inalogy/ssh-connector/releases/tag/v1.2.2
 [1.2.1]: https://github.com/inalogy/ssh-connector/releases/tag/v1.2.1
 [1.2.0]: https://github.com/inalogy/ssh-connector/releases/tag/v1.2.0